lab 5 new requirement added
If a dealer's hand total equals 21 at the start of the game, the dealer has what's called a "natural blackjack." The game is immediately over; neither the player nor the dealer can hit. The dealer automatically wins if the player does not also have a natural blackjack; the game is a tie if the player does have a natural. The game is played out only if the dealer doesn't have a natural blackjack.